Fix installation procedure being stalled
If the configuration directory for the neovim already exist, the
procedure would not try to provision the configuration again. However, in
case if the directory exists only as a result of an incomplete
installation which was tried before, this would give an unwanted result.

From now on the configuration directory is removed before trying to
provision new configuration into it unconditionally.